✨ feat(score): 实现积分与礼品兑换功能
- 新增积分与礼品兑换相关的 service, dal, handler 逻辑
- 定义礼品、兑换记录、积分流水的数据库 schema 和 GORM model
- 实现查询用户积分、获取礼品列表、兑换礼品等 API
- 统一 `score` 相关字段命名和数据类型 (i32 -> i64)
- 移除废弃的校验工具函数和错误码

+++ idl/score.thrift
@@ -8,7 +8,7 @@ struct GetUserScoreReq {
 
 struct GetUserScoreResp {
   required model.BaseResp baseResponse;
-  required i32 score;
+  required i64 score;
 }
 
 // 兑换汽车周边
